+++ Assets/Scripts/Games/SoundEffects.cs
@@ -41,7 +41,7 @@ namespace HeavenStudio.Games
             }
             return sounds;
         }
-        public static void FourBeatCountIn(float beat, float length, int type)
+        public static void FourBeatCountIn(double beat, float length, int type)
         {
             string[] sounds = { "one", "two", "three", "four" };
             sounds = GetCountInSounds(sounds, (CountInType)type);
@@ -55,7 +55,7 @@ namespace HeavenStudio.Games
             }, false);
         }
 
-        public static void EightBeatCountIn(float beat, float length, int type)
+        public static void EightBeatCountIn(double beat, float length, int type)
         {
             string[] sounds = { "one", "two", "one", "two", "three", "four" };
             sounds = GetCountInSounds(sounds, (CountInType)type);
@@ -76,7 +76,7 @@ namespace HeavenStudio.Games
             SoundByte.PlayOneShot("count-ins/cowbell");
         }
 
-        public static void Ready(float beat, float length)
+        public static void Ready(double beat, float length)
         {
             MultiSound.Play(new MultiSound.Sound[]
             {
